Added category for recipe steps
In order for hop-calculation abillity, it was neccessary to implement a new
field in the steps called category. This helps to filter between the different
steps that it's clear where cooking starts and ends. With this information the
hop-calculation will be possible.
Beside that the recipe UI got a little bit of rework.