Calculating the required amount of hop
When creating a new brewing charge the required hop is calculated. Therefore
it calculates the amount of hop to reach the overall IBU which is given in the
recipe. For each sort of hop it looks into the stock, filters all of the same
type and using it's alpha value to calculate the amount. This is done with the
fantastich formula of glenn tinseth. Thanks a lot!